\name{print.tolint}
\alias{print.tolint}
\title{Print a tolerance interval result}
\description{Formatted display of objects returned by
  \code{\link{tolint_oneway}} and \code{\link{tolint_lmer}}.}
\usage{
\method{print}{tolint}(x, digits = 4, ...)
}
\arguments{
  \item{x}{an object of class \code{"tolint"}.}
  \item{digits}{number of digits for the printed limits and estimates.}
  \item{\dots}{ignored.}
}
\value{\code{x}, invisibly.}
\keyword{print}
